gestion

+++ src/Controller/EventController.php
@@ -5,7 +5,9 @@ namespace App\Controller;
 use App\Entity\Event;
 use App\Form\EventType;
 use App\Repository\EventRepository;
+use App\Security\Voter\EventVoter;
 use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -14,28 +16,48 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/evenements')]
 class EventController extends AbstractController
 {
-
-    public function __construct(private readonly EntityManagerInterface $em, private readonly EventRepository $eventRepository)
-    {
+    public function __construct(
+        private readonly EntityManagerInterface $em,
+        private readonly EventRepository $eventRepository,
+        private readonly PaginatorInterface $paginator
+    ) {
     }
 
     #[Route('/', name: 'app_event_index', methods: ['GET'])]
-    public function index(): Response
+    public function index(Request $request): Response
     {
+        if ($this->getUser()) {
+            $query = $this->eventRepository->findAll();
+        } else {
+            $query = $this->eventRepository->findBy(['isPublic' => true]);
+        }
+
+        $pagination = $this->paginator->paginate(
+            $query,
+            $request->query->getInt('page', 1),
+            6
+        );
+
         return $this->render('event/index.html.twig', [
-            'events' => $this->eventRepository->findAll(),
+            'pagination' => $pagination,
         ]);
     }
 
     #[Route('/creer-un-evenement', name: 'app_event_new', methods: ['GET', 'POST'])]
     public function new(Request $request): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('danger', 'Vous devez être connecté pour créer un évènement');
+            return $this->redirectToRoute('app_login');
+        }
+
         $form = $this->createForm(EventType::class);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $event = $form->getData();
             $event->setOwner($this->getUser());
+
             $this->em->persist($event);
             $this->em->flush();
 
@@ -59,7 +81,12 @@ class EventController extends AbstractController
     #[Route('/{id}/modifier', name: 'app_event_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Event $event): Response
     {
-        $form = $this->createForm(EventType::class, $event, ['validation_groups' => ['event']]);
+        if (!$this->isGranted(EventVoter::EDIT, $event)) {
+            $this->addFlash('danger', "Vous n'avez pas la permission de modifier cet évènement.");
+            return $this->redirectToRoute('app_event_show', ['id' => $event->getId()]);
+        }
+
+        $form = $this->createForm(EventType::class, $event);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -71,16 +98,22 @@ class EventController extends AbstractController
 
         return $this->render('event/edit.html.twig', [
             'event' => $event,
-            'form' => $form,
+            'form' => $form->createView(),
         ]);
     }
 
     #[Route('/{id}', name: 'app_event_delete', methods: ['POST'])]
     public function delete(Request $request, Event $event): Response
     {
+        if (!$this->isGranted(EventVoter::DELETE, $event)) {
+            $this->addFlash('danger', "Vous n'avez pas la permission de supprimer cet évènement.");
+            return $this->redirectToRoute('app_event_show', ['id' => $event->getId()]);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$event->getId(), $request->request->get('_token'))) {
             $this->em->remove($event);
             $this->em->flush();
+            $this->addFlash('success', 'Évènement supprimé avec succès');
         }
 
         return $this->redirectToRoute('app_event_index', [], Response::HTTP_SEE_OTHER);
